move helper functions to _internal.py
