+++ pytest_flask/_internal.py
@@ -17,3 +17,19 @@ def deprecated(reason):
         return deprecated_call
 
     return decorator
+
+
+def _rewrite_server_name(server_name, new_port):
+    """Rewrite server port in ``server_name`` with ``new_port`` value."""
+    sep = ":"
+    if sep in server_name:
+        server_name, port = server_name.split(sep, 1)
+    return sep.join((server_name, new_port))
+
+
+def _determine_scope(*, fixture_name, config):
+    return config.getini("live_server_scope")
+
+
+def _make_accept_header(mimetype):
+    return [("Accept", mimetype)]
